feat: maximize artifact reuse between different modes.
As of this writing, it aims to reuse `rmeta` between build and check modes.
(one emits rmeta + rlib, and the other emits rmeta)

This has a caveat that `DefId` in rmeta files might not match between rustc
invocations with different `--emits` values, and that might lead to rustc
rejecting the input rmeta. To avoid failures in rustc, this is currently
guarded by fingerprint to ensure all output files present, if not, rerun.
This make running `check` after `build` safe: `build`'s rmeta files might
have more information and rustc is correct on rejection.