Track the first manifest path as an implicit, fallback root
To allow a package to be unpacked outside of all root directories, such
as `/tmp/my-package` and allow a build to start from
`/tmp/my-package/sub/dir`, Cargo is allowed to traverse parent
directories until it finds its first manifest, and this then becomes a
fallback root directory that will stop Cargo from looking any higher for
a workspace (which could attempt to load `/tmp/Cargo.toml` that may not
be safe).

XXX: Instead of having a special case for the first manifest; maybe it could
be better to instead _always_ add the directory of a manifest as a root
directory if it's not a subdirectory of any existing root.

That would have a mostly-equivalent result to the current behaviour but
_might_ be simpler to document?