Allow artifact target set for build-std
Adds unstable flag `build-std-targets`.

The intended use is for projects that mix a Tier 1/2 target as a host
with a Tier 3 target in an artifact dependency, which is then deployed
at runtime. For instance, x86-64-*-linux and bpfeb-unknown-none, or a
supported host with a custom hardware profile given by json.

In these cases it is unlikely that all std libraries should be built
from scratch. Additionally, the list of std builds are not only those
given as direct targets on the command line but rather discovered by
target specifications on artifact dependencies that are embedded to a
host loader etc.

The previous state discovered the set of std-builds through the target
list of the CLI but never dynamically within the crate graph. The unit
graph would then try to access the std build for kinds that were not
generated, resulting in panics in unit dependency planning.