Recommend `package.rust-version` in `reference/semver.md`.
Adds a recommendation that, to mitigate “possibly breaking” Rust
version requirement increases, they should be declared using
`package.rust-version`.

My main goal in this change is to make it *possible to find* the Rust
version section in this page by searching it, but the most appropriate
way to fit it into this section seemed to me to be making this
recommendation. Declaring (and maintaining) `package.rust-version` is
additional work, but it is a much lesser burden on maintainers than
other items in the same list such as "Provide a large window of support".

Co-authored-by: Ed Page <eopage@gmail.com>