fix(script): Error, rather than sanitize, artfact dir names
Making this an error now gives us the best compatibility story.
A user can workaround this by overriding `package.name`.
Our paths forward include:
- Leave it as-is or at least improve the error message for this case
- Make it so we don't need the error message
- Add back sanitizating the name

Ideally, we make it so we don't need the error message.
The ground work was laid out for this in #16086.
The next step is to move the conflict error from manifest parsing to
compilation so we can check whether target-dir and build-dir overlap to
error.
There are unknowns with this,
including whether the usability is good enough for making this error
conditional on the target-dir and build-dir overlapping.
We may even want to wait until we change the default build-dir.