Fix mdman to not incorrectly strip `<p>` tags
The logic for stripping the `<p>` tag in the option description was not
considering that the description may contain other block elements. This
changes it so that it no longer strips the `<p>` tag. This was done due
to some spacing issues which are better resolved with CSS (and is
resolved in mdbook 0.5).