feat(build): Added dynamic resource limit handling
This commit adds handling to attempt to increase the max file
descriptors if we detect that there is a risk of hitting the limit.

If we cannot increase the max file descriptors, we fall back to coarse
grain locking as that is better than a build crashing due to resource
limits.