docs(ref): Rename DEP_NAME_KEY to DEP_LINKS_KEY
`NAME` is unclear what should go here and the docs have to specifically
call out that it is `package.links` and not `package.name`.
The hope with this rename is to make this even clearer,
especially if we move forward with #3544 and add an alternative to
`DEP_NAME_KEY` that uses the dependency name.