fix(layout): Remove hashes from bins in new layout
There is a desire in #8332 to make our use of `-Cextra-file-name` in `deps/` more
consistent across platforms.
Having it present gets in the way of finding PDBs.
With #15010, we no longer need `-Cextra-file-name` to avoid collisions.
It might still be used by rustc for faster lookups though.
So we can at least adjust all other platforms to be like Windows,
making things consistent

Addresses #8332 if #15010 is stabilized

Part of #15010