feat(report): new command `cargo report sessions`
A new report command `cargo report sessions` is added.
The motivation of this is to provide an access to the build session ID,
so that user can retrieve reports for specific session.
For example, `cargo report timings --id <my-older-build>`.

The command name `sessions` was chosen because

* It is a thing that Cargo "reports to user",
  feeling more reasonable than `cargo report list-runs`
  or `cargo rebuild recent-builds`.
* Discussed with `@Muscraft`,
  and we feel `SessionId` will be a better term in `RunId`.
  While whether a `cargo` invocation is a session is debatable,
  sessions the term is widely used for logging and metrics.
* This is unstable. We can always change its name later.

Part of rust-lang/cargo#15844